Close issue #35
Actually when UpdateStreamAsync is called with high load
and WiredTiger, it is possible that the update of stream
head throws a MongoDuplicateKeyException. In such a
situation we can safely ignore the error, because some
other thread already updated the record and we are not
able to know which is the latest one.